+++ screens/Arrecadaçao/ArrecadacaoTab.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import ArrecadacaoTelaInicial from './ArrecadacaoTelaInicial';
+import TelaInicial from './TelaInicial';
 
 export default function ArrecadacaoTab({ navigation }: { navigation: any }) {
     const Stack = createNativeStackNavigator();
     return (
         <Stack.Navigator>
             <Stack.Screen
-                name="ProdutosEstoque"
-                component={ArrecadacaoTelaInicial}
+                name="ArrecadaoTelaInicial"
+                component={TelaInicial}
                 options={{ headerShown: false }}
             />
         </Stack.Navigator>
