+++ screens/Arrecadaçao/CriarNovaCampanha.tsx
@@ -1,0 +1,8 @@
+import React, { useState } from 'react';
+import { StyleSheet, View, Image } from 'react-native';
+import { Appbar, FAB, Text } from 'react-native-paper';
+import { Button } from 'react-native-paper';
+import { StatusBar } from 'react-native';
+import { ScrollView } from 'react-native';
+import { Dimensions } from 'react-native';
+import { CameraType, useCameraPermissions } from 'expo-camera';
